Scales now support rounding. In addition, the default domain minimum is chosen more intelligently based on the property: 0 is the default for width and height, while the minimum data value is used for left, right, top and bottom.
The value of a def can now (itself) be a function. In this case, the function is
evaluated like normal property functions. This makes def convenient for
constructing a new function for each mark instance, as is needed for scales. See
flowers.html for an example.

Cleanup examples to use new scales.

Added pv.repeat.

git-svn-id: svn+ssh://blit.stanford.edu/usr/graphics/svn/infovis/protovis@613 99c00a00-a864-4cc4-b658-633078b1b672