Add flatten operator, the inverse of the nest operator: it takes a hierarchical map and flattens it into a tabular array. As with nest, you can specify a sequence of keys that specify the attribute names of each level of the hierarchy; an optional map function lets you transform property values.
Job Voyager performance improvements: by representing the data as a hierarchical
map, rather than a tabular array (with lots of redundancies), we reduce the size
of the data file by an order of magnitude (452 KB to 46 KB). In addition, we
populate the map with the missing values so cleanup does not have to be done on
load. This offers another 1.2x performance increase, and the Job Voyager is now
almost twice as fast (~300 ms) as in the previous release!

The alpha argument to pv.rgb and pv.hsl is now optional; if not specified the
color will be fully-opaque.


git-svn-id: svn+ssh://blit.stanford.edu/usr/graphics/svn/infovis/protovis@632 99c00a00-a864-4cc4-b658-633078b1b672