Properties are now evaluated deterministically, in the order they are defined. Well, constants are evaluated before functions, and defs are evaluated before properties, and inherited properties are evaluated first, but other than that... This means you can refer to the value of computed properties reliably, rather than recomputing them.
Reimplement anchors to behave like normal marks. Also assume that properties are
part of the global namespace: if a property is defined on any mark type, it is
implicitly available to all mark types, even if it is not evaluated. For
example, now anchors can define a name property, and that name property can be
queried by a mark that extends an anchor, even if that mark type doesn't
normally have a name property.

As a side-effect of this change, the Job Voyager is another 10% faster. Woot!


git-svn-id: svn+ssh://blit.stanford.edu/usr/graphics/svn/infovis/protovis@689 99c00a00-a864-4cc4-b658-633078b1b672