More improvements to pv.Layout.dendrogram.
Rather than pre-computing the layout positions and then caching them on the
nodes, the positions are computed on-the-fly as transformations of the
normalized (breadth-depth) coordinates. This offers two advantages: (1) if the
parent panel changes size, the layout rescales automatically; (2) the fill
layout can be computed slightly differently than the node-link layout.

The reason the fill layout various from the node-link layout is that the root
node is normally on the edge of the panel (depth=0). In the fill layout, this
would make the root invisible. So we compress the non-root nodes to make room
for the root in the space-filling variant.

Also fix a number of bad sort functions in the dendrogram tests; using the minus
operator trick only works if the fields are numbers. Otherwise, pv.naturalOrder
should be used for correct behavior.