Add partition layout. Remove circle, icicle, radial.
The new pv.Layout.partition replaces the previous icicle (for Cartesian) and
radial (for polar) layouts. It performs a recursive partitioning of a tree.
Also, the dendrogram layout now supports group spacing, meaning it can now
replace the previous circle layout. Add some tests, some of which are now
redundant and may be cleaned up in a future change.