pv.Layout.tree now supports "radial" orientation.
The implementation has been cleaned up such that temporary variables are no
longer exposed after the layout algorithm has run. Rather than setting x- and
y-coordinates directly, the layout algorithm computes the breadth and depth.
(In this case, since the layout algorithm is not space-filling, these are not in
unit space but instead absolute coordinates based on the fixed offsets.) The
breadth and depth can then be transformed using property functions for the node,
link and label marks.

This change also allows you to control space between sibling groups, and
provides a label prototype for convenient node labeling.