Add pv.Constraint.distance to physics simulation.
The physics simulator now uses Position (rather than Velocity) Verlet
integration, which makes it easier to implement distance constraints. In this
commit, a simple distance constraint is added which can be used (for example) to
force particles to be more than a certain distance apart. This will be useful in
implementing a Dorling Cartogram.