Performance improvements to Physics simulation.
Also include new distance constraint that enforces two particles be a fixed
distance apart. This is, in effect, an infinite-tension spring. The names of
these constraints are subject to change (perphaps minDistance and distance?).

The Simulation class now has a stablize() method which can be used to relax
constraints several times before the simulation begins, avoiding spurious
initial velocities. I may change this so this happens automatically when
constraints are added.

Lastly the Simulation class restores the positions of fixed particles in the
event they are inadvertantly moved by a constraint. In a future change I would
like the fixed position to be implemented using a constraint rather than as a
special fixed property, but other API changes need to take place for that to
happen. (Specifically, the pv.Behavior.drag would need a reference to the
simulation to be able to add or remove a constraint.)