Fix bug introduced in previous cast-related change.
The cast function (type) for a property, similar to its name, is global. Thus,
even though fillStyle is defined on Bar and not Mark, the type of fillStyle is
globally a color. This way, if you declare a fillStyle on a generic Mark for the
purpose of inheritance, it continues to be cast to a color.