Add pv.Layout.arc.
This is a simple layout for arc digrams, a simple one-dimensional graph layout.
For example, this may be used to draw arcs along the edge of a matrix
visualization of a graph. This layout also supports radial orientation; at the
moment arcs in the radial layout are drawn as straight lines, but future support
for splines should make this prettier.