Populate SVG attributes more selectively.
Rather than always setting SVG attributes, we now check whether the attribute's
value is the implicit value. If it is, the attribute is removed from the
element, simplifying the generated SVG. In addition, we set a few attributes on
the root SVG elements such that the default (e.g., font) is implicit.

This change is implemented by extending pv.SvgScene.expect to take two
additional optional arguments: a map of SVG attributes, and a map of CSS
properties.