The cursor and title properties now default to null.
Previously, these properties defaulted to the empty string. Since the change to
make property types explicit, it is no longer necessary to convert the title to
a string when setting the title attribute in SVG (because the value of the title
property is already a string, or null). This simplifies the logic and eliminates
the need for a special default for these properties.