First pass at Dorling Cartogram.
This isn't a cartogram per se, because it isn't using a geographic projection
and there isn't geographic data. This part is just a test of using the physics
system and distance constraints.

While repeated relaxation of the distance constraints produces adequate results,
we probably also want to introduce a spring force to pull the circles back to
their initial position. This will require some additional work, because the
current SpringForce is designed to support springs between particles. While we
could simulate the initial positions as particles, it's inefficient; what we
need is a spring pointing to a fixed position.