Improvements to Dorling Cartogram.
Add pv.Constraint.position which constrains each particle to a given position,
computed as a function on each particle. Rename the constraint "spring" to
"link", and the constraint "distance" to "collision"; these still aren't the
best names but I think that makes them less ambiguous.

It's tempting to use pv.Constraint.position as the position constraint for
dragging particles interactively, though in that case the constraint need only
be applied to a single particle rather than all particles. Still to-do.