Formalize required, optional and fixed properties.
The segmented property is considered "fixed", which means that it is only
evaluated on the first instance of the associated Line or Area. Unfortunately,
if the first instance of the Line or Area also happens to be invisible, the
segmented property was not evaluated, and thus all segments disappeared.

Now the segmented property is considered "required", in that it is always
evaluated, even if visible is false. The visible property is similarly required,
though all other properties are "optional", meaning that their evaluation is
skipped if the instance is invisible.

In addition, Lines and Areas now determine on-the-fly which of their properties
are fixed, such that fillStyle and the like are only evaluated on the first
instance for non-segmented marks. Secondary instances get a reference to the
corresponding value from the first instance, rather than re-evaluating the
property, which improves performance while being correct; it is now possible to
use property chaining on a fixed property.