Don't render empty labels.
This fixes a bug where if the value of the text property was null, the first
time a label was rendered it would display "null"; on subsequent renders the
label would be rendered "". (This is due to the different between setting the
nodeValue and using the createTextNode method.)

The simple solution and optimization is to omit the label entirely if the text
property evaluates to false, as is the case with the empty string and null. Also
add a test case for this and the previous commit related to pv.event.