The color `null` is implicitly pv.Color.transparent.
Previously, there were two ways that transparent colors could be represented in
the scene graph: they could be a pv.Color instance with an opacity of 0, or the
value null. This complicated the rendering code as it would need to check for
both values. The code is now simplified by converting any null colors to
pv.Color.transparent in pv.Mark.prototype.buildImplied.