Panels now support transforms!
A transform can now be attached to any panel, causing its children to be offset
and scaled according to the specified transformation matrix. (At the moment, the
panel's own stroke and fill are not transformed, but this is a bug that will be
fixed soon.) The transform also affects the mouse() method; the inverse
transform is used to project the mouse location into the panel's coordinate
space.

This makes it possible to apply pv.Behavior.transform to any panel, allowing the
user to pan and zoom the contained marks! Note: line width and text size are
fixed size, irrespective of transform. This is intentional, though there may be
a need to customize this behavior in the future.