Rendering now preserves scene context.
Previously, rendering would clear the `scene` and `index` properties on marks
such that once rendering finished, any context (such as the context of which
mark instance triggered the event) would be lost. Now the render methods are
more careful about setting and restoring these properties so that this context
is not lost.

This change also moves the current data stack, previously on root.scene.data, to
pv.Mark.stack. The stack is no longer cleared and re-populated when render is
called from an event handler because render now shifts or unshifts data onto the
stack as necessary.