Split pv.Behavior.transform into pan & zoom.
Now that transforms are supported by panels, we can simplify the code by
spliiting the transform behavior into the separate pan & zoom behaviors.

This change also makes a few fixes to pv.Mark.context, which is now used by the
pan behavior to restore the scene context for mousemove and mouseup events: (1)
the specified function is evaluated using the mark as `this`, with the data
stack as `arguments`, similar to property functions and event handlers, (2) the
default index is set, and (3) the current scale is set.