Fixes for scale and pv.Mark.context.
The current scale factor was not being computed correctly if partial redraw was
triggered on a (leaf) mark rather than a panel. In addition, the scale factor
was not being set before calling build, causing properties that were scale-aware
to break. The scale factor is now a bona fide attribute rather than a global,
which is cleaner and also means `this.parent.scale` is now correct.

pv.Mark.context now also wires up child scenes of the context mark, if any. This
allows you to set defs on a child from a panel's event handler. pv.Behavior.drag
now uses pv.Mark.context.