Add pv.Behavior.point.
The "point" behavior finds the closest mark to the mouse location, and then
triggers a "point" event on that mark instance. When the mouse moves away, an
"unpoint" event is triggered. This is similar to mouseover and mouseout events,
except that the mouse need not be precisely over the targeted mark. Also, if
multiple marks overlap, the point event will be dispatched to the closest mark,
rather than the one that happens to be on top.

The point behavior can be configured with a minimum distance, so that if the
closest mark is far away, no event will be dispatched. You can also collapse a
dimension in x or y, which is useful in plots with a dependent and independent
variable (e.g., time series plots).

There remains a bug with partial redraw when property chaining is used. In this
case, the chain target can have an index set (e.g., the chain target handled the
event and then called render on the root), so when the property is evaluated it
will query target.scene[target.index] instead of target.scene[this.index].