Fixes for property chaining and partial rendering.
The new mark instance() method returns the current instance in the scene graph.
If the index is not set, the default index is used; if no default index is
available, the last index is used. Similarly, if the scene is not set, the last
scene in the enclosing panel is used, and so on.

This method allows you to query property values even if an index is not set.
This is particularly nice in the case of singleton marks (where there is only
one valid index), but also means inheritance and property chaining can now be
used across panels.

This commit also fixes a recently-introduced bug where we would inadvertantly
render from the root, rather than rendering just the requested marks.