Layouts now use properties for configuration.
Previously, layout configuration was static. If, for example, you used a layout
within a panel, you could not vary the layout across panels. Now all layout
configuration parameters are computed as properties, rather than being set to
constant values.

This means that you can pass in property functions to compute the layout
configuration (and pass in data) dynamically, allowing layout configuration to
vary from panel to panel, chain from other properties, etc. So, it is now
trivial to make a small multiples display of layouts with varying data or
configurations. (Previously, this would require code duplication or a loop.)

This commit is a work in-progress; so far only stack layout has been updated to
use the new framework.