Grid layout now uses properties.
You can specify the data associated with a grid layout using a property, such
that the data can be dereferenced from an enclosing panel. Similarly you can now
specify the grid layout in terms of number of rows and columns, which implicitly
sets the data to replicate the enclosing panel's data.