Fix for def getter-setters.
Previously, properties and defs had different paths for getting and setting
values. This meant that getting or setting the value of a def after render
failed, since the scene was undefined. Now, if the scene is set, a def
getter-setter will use the associated scene defs; otherwise, it will fall back
to setting the def's initial value (for a subsequent render), or getting the
value from the default scene.