Improvements to force-directed layout.
Nodes and links are no longer wrapped on-input (with the exception of primitive
nodes, such as strings), which makes the data transformation less confusing and
more convenient. In particular, it is now easy to define links as a function of
the nodes using conventional property chaining.

Rather than iterating indefinitely, the layout now tracks the speed of the
slowest particle in the simulation; if this speed is below some threshold, the
simulation (and rendering) is paused, saving CPU cycles. Simulation
automatically resumes if a particle is dragged interactively, as determined by
querying the `fixed` attribute.

Meanwhile, the `interactive` property has been removed; if a number of
iterations has been specified, this implies the layout is non-interactive and
will be computed before the initial render.

A bound constraint may be added to enforce that the particles do not exit the
containing panel. This is not enabled by default, since it is not necessary in
conjunction with panning & zooming. Currently, the bounds are padded by a fixed
6 pixels, but this could be made smarter in a future change by querying the node
radius and lineWidth.

The random walk to set initial particle positions is now biased by the aspect
ratio of the containing panel, so that the layout is more likely to converge on
a matching orientation.