Add hierarchy layout; pack now extends hierarchy.
A hierarchy layout is a special case of a network layout where the links are
computed implicitly from the nodes' parents. This facilitates code reuse and
consistent behavior.

A side-effect of this change is that the data specified for pack layouts is now
an array of nodes (the same format as that used by force-directed, arc, and
other network layouts). It is easy to convert from a map to an array of nodes
using pv.Dom.