Better rendering of discontinuous areas.
Previously, if the height of an area was zero for several points, a stroke would
still be drawn at the bottom position. Now we detect these discontinuous gaps in
the area, and render each section using a separate subpath, such that the stroke
goes around each section but not in-between. This improves the rendering of
visualizations such as stacked area graphs where the data may be zero for some
points.