Inject stack layout with awesomeness.
The `offset` property now takes a string rather than a function, and implements
two standard algorithms: silohouette and wiggle. The first, also known as
ThemeRiver, centers the stacks around the middle of the panel. The second,
described in Byron & Wattenberg's "Stacked Graphs -- Geometry & Aesthetics"
paper, minimizes the weighted change in slope (the "wiggle"). The default offset
is "zero".

A new `order` property adopts a further recommendation by Byron & Wattenberg; by
specifying the value "inside-out", the layout will first order the series by the
occurrence of their maximum value, and then interleave the series to provide
equal weighting on the top and bottom. The "inside-out" order is highly
recommended in conjunction with the "wiggle" offset, as without it the wiggle
algorithm can drift.