Unlock def values if the associated property is redefined.
A def can be "locked" by setting its value within a mark context, such as from
within an event handler; this allows the locked value to be used for that mark
instance, rather than re-evaluating the def initializer.

However, once a def value was locked, there was no way to unlock it except from
within the same context. Re-assigning the def initializer (i.e., setting the def
value from a global context) would not associated unlock the def values.

Now we track a sequence id for assigned properties and def values. This way, if
a def is locked, and then subsequently the property is redefined, we know to
reevaluate the def rather than using the old locked value.

Network layouts implicitly lock the values of the `nodes` and `links`
properties, such that the layout can be re-rendered quickly without restarting
the layout. (Consider the case of force-directed layout!) Now if these
properties are explicitly reassigned, the layout will discard the old nodes and
links and restart.