Fix a bug in stack layout with right or bottom.
After bind is called, any undefined properties are bound with the null value. So
it looked like the left and top properties were always defined, even though they
weren't. Now we look to see if the left or top property is bound to null, and if
so, use the alternative.

For maximum compatibility, we should probably invoke both properties for each
sample, and then compute the implied value as is done in Mark.buildImplied. But
it would be rather unusual to define a stacked graph where some of the samples
are defined using left and others are defined using right, so it seems
reasonable to make this restriction. And at any rate, the x-property is only
needed by offset("wiggle") to calculate slope.