Linear scale now supports customizable tick count.
By passing an argument to the ticks() method, the user can now specify the
desired number of ticks. The scale determines a step interval such that the
returned array of ticks is as close as possible to the desired number of ticks.
Note: this is only supported on linear scale for numeric values.