Fix a bug in hierarchy layout's implicit links.
The implicit links property needs to be bound to the layout; otherwise, it will
simply be re-added on subsequent renders, causing it be re-evaluated even if the
nodes have not changed. It would be nice if we could define this using the
standard defaults mechanism, but unfortuantely it depends on the nodes property
and so needs to be registered last.