Add tests for non-clipped stacked area anchors.
When adding an anchored mark to a stacked area directly, the anchored mark and
the areas are drawn interleaved; this frequently causes clipping as one series'
area paints over an earier series' anchored mark.

While reversing the render order of the series (using `reverse`) can alleviate
some of these issues, a better solution is to draw the anchored marks on a
separate plane above the area. But since we don't support arbtirary z-order,
then the problem becomes: how to access the scene state of the area, since
anchors cannot be added across panels?

Fortunately, there is a simple solution: create a transparent copy of the
stacked areas, extending the original. Add the anchored marks to the overlay.
Since the overlay is transparent no clipping will occur. Another advantage of
putting the anchored labels in an overlay is that they can be redrawn very
quickly, without recomputing or redrawing the stacked layout!

This commit includes two examples: one that renders anchored labels for all
series, and one that renders an anchored label for only the closest series.