Fix pv.parse for IE.
The problem with .innerHTML is that it automatically escapes the
contents e.g. '<' will become '&gt;'.  It turns out `.text` is the
equivalent property for IE, and it also works in Firefox and
Webkit-based browsers.  I've tested this in IE6, IE7, IE8 and IE9
(Platform Preview), as well as Google Chrome, Safari, Firefox and Opera.