Treemap improvements.
Treemaps now support two modes: "squarify" for squarified treemaps (as before),
and "slice-and-dice" for the simpler algorithm that alternates between
horizontal and vertical cuts at different hierarchy levels. In addition,
treemaps now support three orders: "ascending", "descending", and null (for
unsorted). The default is "ascending" as before.

This commit also fixes an occasional bug with rounding.