Treemap improvements: hierarchical layouts.
Inspired by Slingsby et al.'s "Configuring Hierarchical Layouts to Address
Research Questions" (HiVE), I've made it easy to create hierarchies of treemap
layouts. This is implemented by creating a separate treemap layout for each
level of the hierarchy.

Treemaps now support two additional modes: "slice" for vertical slices, and
"dice" for horizontal slices. This allows control over the hierarchical layout,
in comparison to "slice-and-dice" which determines slice orientation by node
depth.

Adding a pv.Layout to a pv.Layout.Treemap now implicitly adds the layout to a
panel first. In addition, the panel is configured with no stroke and fill, and
is set to only be visible for leaf nodes in the treemap. (These properties can
of course be changed by the user.)

A new `pv.nodes` operator allow easy conversion from an array of objects to a
one-level hierarchy of nodes, with the `nodeValue` attribute pointing to the
corresponding value in the original array.