Fix for wheelDelta in latest versions of WebKit.
It looks like WebKit recently changed the wheelDelta value to be multipled by
120 to match Windows. Unfortunately, this leads to different behavior between
Safari 4.0.5 and WebKit nightly builds. Furthermore, it practice the change
appeared to have affected the wheelDelta by a factor of ~48 rather than 120.

https://bugs.webkit.org/show_bug.cgi?id=35566
http://gitorious.org/webkit/webkit/commit/8645162efca59894c767ff3053525a0a9200fdca

I've added a `speed` argument to the zoom behavior constructor which would allow
the user to override the default zoom speed.