The flatten operator can now use a leaf function.
Previously the flatten operator could only be used with a specific set of keys.
Now a leaf function can be specified and the keys will be determined dynamically
and stored in a `keys` array. This makes it easy to use the flatten operator as
the inverse of the tree operator.