Add eccentricity property to polar lines.
The eccentricity property allows you to control the straightness of a
polar-interpolated line segment: the default value of 0 draws a circular arc,
while a value of 1 draws a straight line. A value inbetween 0 and 1 draws an
elliptical arc with the given eccentricity.

This commit also introduces a new "polar-reverse" interpolation for lines, which
is the same as "polar" except it draws the arc in a counterclockwise (negative
angle) direction, rather than the default clockwise (positive angle) direction.
In addition, the interpolate property is no longer fixed, so it can vary across
line segments.

The eccentricity property is needed by the rollup layout to indicate
directionality of edges, and to avoid ambiguities when links intersect nodes.