Integrate with vads-protovis.
Mostly this is minor reformatting and cleanup. There are a few API changes.

SPLINES

This commit includes Vad's implementation of "basis" and "cardinal"
interpolation for lines. Note that these interpolation methods are not yet
supported for areas, nor segmented lines, nor segmented areas. Support for these
is forthcoming.

The "interpolate" property is once again fixed for non-segmented lines and
areas. Although there are some contexts in which it may be valid to change the
interpolate property across lines, that is certainly not true for splines and it
seems confusing to support it intermittently; better to just make it fixed. The
area-zero test is updated accordingly.

GEO SCALES

For convenience, geo scales now export `x` and `y` scale methods, which can be
assigned to the `left` and `top` properties. Geo scales now also have two ticks
methods: `ticks.lat` and `ticks.lng`. The `projection` method now behaves
correctly when passed a single string argument (e.g., "mercator"). The `range`
method can now accept two number arguments, which specify the maximum values of
the range while the minimum values are implicitly zero.

Note: geo domains are different than those of quantitative scales because the
projection is two-dimensional and non-affine. This means that a geo domain
cannot be represented as either a one-dimensional (span) or two-dimensional
(rect) range. Instead the domain is a set of points that when projected define a
domain in normalized ([-1,1]) space. Thus, the default domain is now the empty
array, which resets the normalized domain to [1,-1] in both dimensions. This
means that the user need not specify a default domain to project the entire
world!

In addition, the arguments to the pv.Geo.scale constructor are interpreted as
the projection, rather than the domain. Vad initially implemented it this way,
and I thought it would be more consistent with other scales if the arguments
were the domain. However, now that I understand how different the domain is for
geo scales, it seems more sensible (and convenient!) to have the arguments
forwarded to the projection method.

The pv.Geo.projection interface has been changed slightly; the new methods are
`project` and `invert`. The names are intended to match pv.Scale (which also has
an `invert` method). The geo-related code now properly uses the `lng` attribute,
rather than `lon`, to match the pv.Geo.LatLng interface. The mercator projection
is undefined when Math.abs(lat) > 85; we now clamp the projected values to -1
and 1 rather than returning Infinity or -Infinity, which causes the browser's
SVG renderer to throw a fatal error.

The pv.Geo.radians and pv.Geo.degrees methods have been moved to pv, living in
src/data/Numbers.js.

PHYSICS

I deleted the pv.Constraint.link, for the time-being. I'll probably add it back
in the future, but it's not used by any examples and the design still needs to
be determined.

EXAMPLES

I restructured the geo scale examples, breaking them up into smaller examples
suitable for the tests directory. The countries.js data file now converts itself
to this representation on load, eliminating extra conversion on render (as well
as duplicate code).

I deleted the dorling-us example, but this is only temporary and I plan on
adding it back once some additional work has been done. Similarly, I eliminated
all code related to Proj4js; when we have an example of Proj4js integration
cleaned up and ready to go it can live in examples/proj4js.