Simplify grid and horizon layouts using panels.
Rather than extending Mark (and Layout) and doing some behind-the-scenes
shenanigans to produce the desired effect, the grid and horizon layouts now
extend Panel. They use no special hacks behind the scenes to accomplish the
layout, using only standard properties to implement the layout.

This simplifies the layout code, while also making the layouts behave more like
standard panels. And, they're more functional to boot! For example, before to
nest grid layouts, we might say:

  vis.add(pv.Layout.Grid)
      .data([[ ... row-major data ... ]])
    .add(pv.Panel)
    .add(pv.Layout.Grid)
      .data(pv.transpose([[ ... column-major data ... ]])))

Now we can say:

  vis.add(pv.Layout.Grid)
      .rows([[ ... row-major data ... ]])
    .add(pv.Layout.Grid)
      .cols([[ ... column-major data ... ]])

Because the layout is a panel, we can add directly to it (and nest layouts). We
define the data indirectly using the "rows" or "cols" property, rather than
subtly changing the meaning of the data property behind-the-scenes. Thus, you
can now override the data property with a grid layout, if desired.

Similarly, with the horizon layout, we now add areas to the `band` mark, rather
than adding them directly to the layout. (No more behind-the-scenes implicit
reparenting and inheritance!) And, the horizon layout no longer changes the
meaning of the "fillStyle" property, instead using a new explicit property,
"backgroundStyle".

To be sure, layouts may yet be opaque, as the way that they internally define
properties on the panel or child marks will always be partly an implementation
detail. But the goal here is that layouts should use only the mechanisms
typically available to users, rather than changing the internal behavior of
property evaluation or rendering. (At least, where possible--some layouts may be
special enough to require exemptions.) This way, if the user wants to implement
their own layout, they don't have to learn all new abstractions, and can use
what they've learned implementing custom visualizations.