Simplify stack layout using panels.
In the same vein as commit b4f6884, stack layout now extends panel. Rather than
performing a complex manipulation of the bind and build process, the stack
layout uses new psuedo-properties `step` and `size` to compute the layout. The
data is now defined using the `layers` property, which allows the stack layout
to be computed as a side-effect of the inherited data property.

I have temporarily removed support for labels with stack layouts. Although that
would be possible, I'd like to have that supported through improvements in the
anchor system.