Allow anchors across panels.
Previously, using anchors across panels would fail because Protovis did not know
which instances of the anchor target to reference. (The code would crash because
`target.scene is undefined.`) Now we support mirrored descent, which means that
if the anchor and its target are in different panels, we find the target
instances using the anchor's parent indexes.

For example, if you have a stacked area:

  var area = vis.add(pv.Panel)
      .data(data)
    .add(pv.Area)
      ...

You can now easily add labels in an overlay:

  vis.add(pv.Panel)
      .extend(area.parent)
    .add(pv.Label)
      .extend(area.anchor("center"))
      ...

These labels will not be clipped by the areas since they exist in a separate
panel.