Fix a bug in panel anchors.
This changes the implementation of panel anchors slightly, such that the special
behavior for panel anchors is limited to when marks are added directly to the
panel anchor. In this case, only one instance of the mark will be generated per
panel, since the panel will serve to replicate the mark.