Anchors now chain all properties.
This allows anchors to be used to replicate marks statically, without
re-evaluating their properties. Panels did this previously, in a sense, except
they only supported the positional properties (top, left, right, bottom) and
visible. For example, given a stacked `area`:

  vis.add(pv.Panel)
      .extend(area.parent.anchor())
    .add(pv.Area)
      .extend(area.anchor())

creates an exact copy of the area, including the computed fillStyle and the
like.