Rewrite select behavior.
The select behavior is applied to a panel to allow the user to specify the
selected region by dragging; this is also know an "brushing". In addition, the
drag behavior can now be used with rectangular regions, expressed by {x, y, dx,
dy}. The dragged region will be bounded by the parent panel.