Simplify network layouts using panels.
This makes network layouts less magical, simplifying the implementation and
hopefully making them more comprehensible for users. A new "cache" property on
network layouts specifies whether the layout should be cached. The default is
true; caching can be temporarily disabled within an event handler to force
redraw. Rename angle to midAngle to avoid confusion with wedge's angle property.

Treemap layout now has a separate `leaf` mark which is only displayed on leaf
nodes, and does not have any fill or stroke. This simplifies the implementation
of nested (hierarchical) treemaps.

The rollup layout no longer uses the standard positional properties to layout
the nodes, since now that the rollup layout is a panel, we can't multiply the
panel and still render the rolled up nodes and links easily. Instead, the
psuedo-properties x and y are evaluated, similar to stack layouts.