Stack layout now has `values` function.
The values function is equivalent to the layer's data function, meaning that the
layers function now need not return a two-dimensional array. Instead, like data
properties, it returns any array, and the values function is used to dereference
the values for each layer. The default value is the identity function.