Fix a bug in network layout caching.
The network layout now stores a flag on the scene, so that it knows whether the
given scene instance has been initialized (and therefore is cached). Also, the
cache property is unlocked, such that if the cache property is set to false
within an event handler, it is automatically reset using its initializer on
a subsequent render. Thus, if an event triggers some change that causes the
network layout to clear the cache, the network layout will continue to be cached
afterwards.