Fix point behavior for marks with width and height.
The point behavior now uses the mark centroid, rather than the top-left point,
to compute distance from the cursor. This functionality is not yet implemented
for wedges, but pointing can be used with wedges in conjunction with an anchor
serving as the centroid, as shown in the included test.

In the future, it would be better for the point behavior to compute the true
distance from the mouse to the closest point on the perimeter of the mark,
taking into consideration the true shape of the mark. But the current approach
is simple, fast and sufficient for now.