Use the node, not value, to size hierarchical layouts.
Previously the size psuedo-property would get passed the node's nodeValue
attribute. In general, this is what is needed to compute the size; however, it
is inconsistent with other properties which get passed the nodes, since the
nodes are used as the underlying data elements. For consistency, then, we pass
the node rather than node.nodeValue to the size function.