Remove special anchor logic for width & height.
This logic was designed to facilitate stacking bars, but this is no longer a
good practice given the much more powerful stack layout. Furthermore, it
prevents us from doing simple things like anchoring a rule on a bar. Eliminating
the special case makes anchors more consistent and easier to understand.