Render more selectively when dragging.
Originally, the drag behavior only re-rendered the mark instance being dragged.
However, in some situations, dragging one mark requires that related marks also
be re-rendered. For example, with a network layout, it is desirable to re-render
both the node being dragged and any connected links.

In general it is not possible for the drag behavior to know which marks need to
be re-rendered, as it does not know the underlying dependencies. However, the
recently-introduced drag events provides a good opportunity for users to specify
which marks should be re-rendered upon drag events. Accordingly, we can restore
the original default behavior, such that only the mark instance being dragged is
automatically re-rendered. Any other marks can be re-rendered by registering a
corresponding "drag" event handler.

For convenience, the Mark `event` method now accepts constants, which are
wrapped using pv.functor. This allows a mark to be registered directly as an
event handler, such that this mark is re-rendered when the event is handled.